RENAME: types 파일 폴더 이동
- 상수 보관용 폴더로 이동
